Phase 11: Consolidate utils and cleanup
- Move deployments/utils/* → utils/
- Update all utils imports
- Remove empty deployments/ and configs/ directories
