+++ python/ray/llm/_internal/serve/engines/vllm/vllm_engine.py
@@ -31,7 +31,7 @@ from ray.llm._internal.serve.core.engine.protocol import LLMEngine
 from ray.llm._internal.serve.engines.vllm.vllm_models import (
     VLLMEngineConfig,
 )
-from ray.llm._internal.serve.deployments.utils.node_initialization_utils import (
+from ray.llm._internal.serve.utils.node_initialization_utils import (
     initialize_node,
 )
 from ray.llm._internal.serve.observability.logging import get_logger
