+++ python/ray/llm/tests/serve/cpu/deployments/utils/test_batcher.py
@@ -7,7 +7,7 @@ import numpy as np
 import pytest
 
 from ray.llm._internal.serve.constants import MODEL_RESPONSE_BATCH_TIMEOUT_MS
-from ray.llm._internal.serve.deployments.utils.batcher import Batcher
+from ray.llm._internal.serve.utils.batcher import Batcher
 
 TEXT_VALUE = "foo"
 FINAL_TEXT_VALUE = "bar"
